requirements.txt
asgiref==3.7.2
Django==3.2.10
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
PyMySQL==1.1.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
